requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
openpyxl>=3.1.0
python-multipart>=0.0.6
